refactor file directory to match structure of the project
