+++ app/src/main/res/layout/activity_signup.xml
@@ -10,7 +10,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".ui.Signup">
+    tools:context=".SignupActivity">
 
     <View
         android:id="@+id/view2"
